api/requirements.txt
Flask>=3.0
googletrans-py